Move files to shared partition
