+++ src/mscorlib/shared/System/Int16.cs
@@ -6,82 +6,78 @@
 **
 **
 **
-** Purpose: 
+** Purpose: This class will encapsulate a short and provide an
+**          Object representation of it.
 **
-**
+** 
 ===========================================================*/
 
-using System.Globalization;
-using System;
-using System.Runtime.InteropServices;
 using System.Diagnostics.Contracts;
+using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 
 namespace System
 {
-    // A place holder class for signed bytes.
     [Serializable]
-    [CLSCompliant(false), System.Runtime.InteropServices.StructLayout(LayoutKind.Sequential)]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
-    public struct SByte : IComparable, IFormattable, IConvertible
-        , IComparable<SByte>, IEquatable<SByte>
+    [StructLayout(LayoutKind.Sequential)]
+    [TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
+    public struct Int16 : IComparable, IConvertible, IFormattable, IComparable<Int16>, IEquatable<Int16>
     {
-        private sbyte m_value; // Do not rename (binary serialization)
+        private short m_value; // Do not rename (binary serialization)
 
-        // The maximum value that a Byte may represent: 127.
-        public const sbyte MaxValue = (sbyte)0x7F;
-
-        // The minimum value that a Byte may represent: -128.
-        public const sbyte MinValue = unchecked((sbyte)0x80);
-
+        public const short MaxValue = (short)0x7FFF;
+        public const short MinValue = unchecked((short)0x8000);
 
         // Compares this object to another object, returning an integer that
         // indicates the relationship. 
         // Returns a value less than zero if this  object
         // null is considered to be less than any instance.
-        // If object is not of type SByte, this method throws an ArgumentException.
+        // If object is not of type Int16, this method throws an ArgumentException.
         // 
-        public int CompareTo(Object obj)
+        public int CompareTo(Object value)
         {
-            if (obj == null)
+            if (value == null)
             {
                 return 1;
             }
-            if (!(obj is SByte))
+
+            if (value is Int16)
             {
-                throw new ArgumentException(SR.Arg_MustBeSByte);
+                return m_value - ((Int16)value).m_value;
             }
-            return m_value - ((SByte)obj).m_value;
+
+            throw new ArgumentException(SR.Arg_MustBeInt16);
         }
 
-        public int CompareTo(SByte value)
+        public int CompareTo(Int16 value)
         {
             return m_value - value;
         }
 
-        // Determines whether two Byte objects are equal.
         public override bool Equals(Object obj)
         {
-            if (!(obj is SByte))
+            if (!(obj is Int16))
             {
                 return false;
             }
-            return m_value == ((SByte)obj).m_value;
+            return m_value == ((Int16)obj).m_value;
         }
 
-        [System.Runtime.Versioning.NonVersionable]
-        public bool Equals(SByte obj)
+        [NonVersionable]
+        public bool Equals(Int16 obj)
         {
             return m_value == obj;
         }
 
-        // Gets a hash code for this instance.
+        // Returns a HashCode for the Int16
         public override int GetHashCode()
         {
-            return ((int)m_value ^ (int)m_value << 8);
+            return ((int)((ushort)m_value) | (((int)m_value) << 16));
         }
 
 
-        // Provides a string representation of a byte.
         public override String ToString()
         {
             Contract.Ensures(Contract.Result<String>() != null);
@@ -112,43 +108,35 @@ namespace System
 
             if (m_value < 0 && format != null && format.Length > 0 && (format[0] == 'X' || format[0] == 'x'))
             {
-                uint temp = (uint)(m_value & 0x000000FF);
+                uint temp = (uint)(m_value & 0x0000FFFF);
                 return Number.FormatUInt32(temp, format, info);
             }
             return Number.FormatInt32(m_value, format, info);
         }
 
-        [CLSCompliant(false)]
-        public static sbyte Parse(String s)
+        public static short Parse(String s)
         {
             return Parse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo);
         }
 
-        [CLSCompliant(false)]
-        public static sbyte Parse(String s, NumberStyles style)
+        public static short Parse(String s, NumberStyles style)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return Parse(s, style, NumberFormatInfo.CurrentInfo);
         }
 
-        [CLSCompliant(false)]
-        public static sbyte Parse(String s, IFormatProvider provider)
+        public static short Parse(String s, IFormatProvider provider)
         {
             return Parse(s, NumberStyles.Integer, NumberFormatInfo.GetInstance(provider));
         }
 
-        // Parses a signed byte from a String in the given style.  If
-        // a NumberFormatInfo isn't specified, the current culture's 
-        // NumberFormatInfo is assumed.
-        // 
-        [CLSCompliant(false)]
-        public static sbyte Parse(String s, NumberStyles style, IFormatProvider provider)
+        public static short Parse(String s, NumberStyles style, IFormatProvider provider)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return Parse(s, style, NumberFormatInfo.GetInstance(provider));
         }
 
-        private static sbyte Parse(String s, NumberStyles style, NumberFormatInfo info)
+        private static short Parse(String s, NumberStyles style, NumberFormatInfo info)
         {
             int i = 0;
             try
@@ -157,36 +145,36 @@ namespace System
             }
             catch (OverflowException e)
             {
-                throw new OverflowException(SR.Overflow_SByte, e);
+                throw new OverflowException(SR.Overflow_Int16, e);
             }
 
+            // We need this check here since we don't allow signs to specified in hex numbers. So we fixup the result
+            // for negative numbers
             if ((style & NumberStyles.AllowHexSpecifier) != 0)
             { // We are parsing a hexadecimal number
-                if ((i < 0) || i > Byte.MaxValue)
+                if ((i < 0) || (i > UInt16.MaxValue))
                 {
-                    throw new OverflowException(SR.Overflow_SByte);
+                    throw new OverflowException(SR.Overflow_Int16);
                 }
-                return (sbyte)i;
+                return (short)i;
             }
 
-            if (i < MinValue || i > MaxValue) throw new OverflowException(SR.Overflow_SByte);
-            return (sbyte)i;
+            if (i < MinValue || i > MaxValue) throw new OverflowException(SR.Overflow_Int16);
+            return (short)i;
         }
 
-        [CLSCompliant(false)]
-        public static bool TryParse(String s, out SByte result)
+        public static bool TryParse(String s, out Int16 result)
         {
             return TryParse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo, out result);
         }
 
-        [CLSCompliant(false)]
-        public static bool TryParse(String s, NumberStyles style, IFormatProvider provider, out SByte result)
+        public static bool TryParse(String s, NumberStyles style, IFormatProvider provider, out Int16 result)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return TryParse(s, style, NumberFormatInfo.GetInstance(provider), out result);
         }
 
-        private static bool TryParse(String s, NumberStyles style, NumberFormatInfo info, out SByte result)
+        private static bool TryParse(String s, NumberStyles style, NumberFormatInfo info, out Int16 result)
         {
             result = 0;
             int i;
@@ -195,13 +183,15 @@ namespace System
                 return false;
             }
 
+            // We need this check here since we don't allow signs to specified in hex numbers. So we fixup the result
+            // for negative numbers
             if ((style & NumberStyles.AllowHexSpecifier) != 0)
             { // We are parsing a hexadecimal number
-                if ((i < 0) || i > Byte.MaxValue)
+                if ((i < 0) || i > UInt16.MaxValue)
                 {
                     return false;
                 }
-                result = (sbyte)i;
+                result = (Int16)i;
                 return true;
             }
 
@@ -209,7 +199,7 @@ namespace System
             {
                 return false;
             }
-            result = (sbyte)i;
+            result = (Int16)i;
             return true;
         }
 
@@ -219,7 +209,7 @@ namespace System
 
         public TypeCode GetTypeCode()
         {
-            return TypeCode.SByte;
+            return TypeCode.Int16;
         }
 
 
@@ -235,7 +225,7 @@ namespace System
 
         sbyte IConvertible.ToSByte(IFormatProvider provider)
         {
-            return m_value;
+            return Convert.ToSByte(m_value);
         }
 
         byte IConvertible.ToByte(IFormatProvider provider)
@@ -245,7 +235,7 @@ namespace System
 
         short IConvertible.ToInt16(IFormatProvider provider)
         {
-            return Convert.ToInt16(m_value);
+            return m_value;
         }
 
         ushort IConvertible.ToUInt16(IFormatProvider provider)
@@ -255,7 +245,7 @@ namespace System
 
         int IConvertible.ToInt32(IFormatProvider provider)
         {
-            return m_value;
+            return Convert.ToInt32(m_value);
         }
 
         uint IConvertible.ToUInt32(IFormatProvider provider)
@@ -290,7 +280,7 @@ namespace System
 
         DateTime IConvertible.ToDateTime(IFormatProvider provider)
         {
-            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "SByte", "DateTime"));
+            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "Int16", "DateTime"));
         }
 
         Object IConvertible.ToType(Type type, IFormatProvider provider)
