+++ src/mscorlib/shared/System/SByte.cs
@@ -2,83 +2,76 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-/*============================================================
-**
-**
-**
-** Purpose: This class will encapsulate a short and provide an
-**          Object representation of it.
-**
-** 
-===========================================================*/
-
-
-using System;
-using System.Globalization;
-using System.Runtime.InteropServices;
 using System.Diagnostics.Contracts;
+using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 
 namespace System
 {
     [Serializable]
-    [System.Runtime.InteropServices.StructLayout(LayoutKind.Sequential)]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
-    public struct Int16 : IComparable, IFormattable, IConvertible
-            , IComparable<Int16>, IEquatable<Int16>
+    [CLSCompliant(false)]    [StructLayout(LayoutKind.Sequential)]
+    [TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
+    public struct SByte : IComparable, IConvertible, IFormattable, IComparable<SByte>, IEquatable<SByte>
     {
-        private short m_value; // Do not rename (binary serialization)
+        private sbyte m_value; // Do not rename (binary serialization)
 
-        public const short MaxValue = (short)0x7FFF;
-        public const short MinValue = unchecked((short)0x8000);
+        // The maximum value that a Byte may represent: 127.
+        public const sbyte MaxValue = (sbyte)0x7F;
+
+        // The minimum value that a Byte may represent: -128.
+        public const sbyte MinValue = unchecked((sbyte)0x80);
+
 
         // Compares this object to another object, returning an integer that
         // indicates the relationship. 
         // Returns a value less than zero if this  object
         // null is considered to be less than any instance.
-        // If object is not of type Int16, this method throws an ArgumentException.
+        // If object is not of type SByte, this method throws an ArgumentException.
         // 
-        public int CompareTo(Object value)
+        public int CompareTo(Object obj)
         {
-            if (value == null)
+            if (obj == null)
             {
                 return 1;
             }
-
-            if (value is Int16)
+            if (!(obj is SByte))
             {
-                return m_value - ((Int16)value).m_value;
+                throw new ArgumentException(SR.Arg_MustBeSByte);
             }
-
-            throw new ArgumentException(SR.Arg_MustBeInt16);
+            return m_value - ((SByte)obj).m_value;
         }
 
-        public int CompareTo(Int16 value)
+        public int CompareTo(SByte value)
         {
             return m_value - value;
         }
 
+        // Determines whether two Byte objects are equal.
         public override bool Equals(Object obj)
         {
-            if (!(obj is Int16))
+            if (!(obj is SByte))
             {
                 return false;
             }
-            return m_value == ((Int16)obj).m_value;
+            return m_value == ((SByte)obj).m_value;
         }
 
-        [System.Runtime.Versioning.NonVersionable]
-        public bool Equals(Int16 obj)
+        [NonVersionable]
+        public bool Equals(SByte obj)
         {
             return m_value == obj;
         }
 
-        // Returns a HashCode for the Int16
+        // Gets a hash code for this instance.
         public override int GetHashCode()
         {
-            return ((int)((ushort)m_value) | (((int)m_value) << 16));
+            return ((int)m_value ^ (int)m_value << 8);
         }
 
 
+        // Provides a string representation of a byte.
         public override String ToString()
         {
             Contract.Ensures(Contract.Result<String>() != null);
@@ -109,35 +102,43 @@ namespace System
 
             if (m_value < 0 && format != null && format.Length > 0 && (format[0] == 'X' || format[0] == 'x'))
             {
-                uint temp = (uint)(m_value & 0x0000FFFF);
+                uint temp = (uint)(m_value & 0x000000FF);
                 return Number.FormatUInt32(temp, format, info);
             }
             return Number.FormatInt32(m_value, format, info);
         }
 
-        public static short Parse(String s)
+        [CLSCompliant(false)]
+        public static sbyte Parse(String s)
         {
             return Parse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo);
         }
 
-        public static short Parse(String s, NumberStyles style)
+        [CLSCompliant(false)]
+        public static sbyte Parse(String s, NumberStyles style)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return Parse(s, style, NumberFormatInfo.CurrentInfo);
         }
 
-        public static short Parse(String s, IFormatProvider provider)
+        [CLSCompliant(false)]
+        public static sbyte Parse(String s, IFormatProvider provider)
         {
             return Parse(s, NumberStyles.Integer, NumberFormatInfo.GetInstance(provider));
         }
 
-        public static short Parse(String s, NumberStyles style, IFormatProvider provider)
+        // Parses a signed byte from a String in the given style.  If
+        // a NumberFormatInfo isn't specified, the current culture's 
+        // NumberFormatInfo is assumed.
+        // 
+        [CLSCompliant(false)]
+        public static sbyte Parse(String s, NumberStyles style, IFormatProvider provider)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return Parse(s, style, NumberFormatInfo.GetInstance(provider));
         }
 
-        private static short Parse(String s, NumberStyles style, NumberFormatInfo info)
+        private static sbyte Parse(String s, NumberStyles style, NumberFormatInfo info)
         {
             int i = 0;
             try
@@ -146,36 +147,36 @@ namespace System
             }
             catch (OverflowException e)
             {
-                throw new OverflowException(SR.Overflow_Int16, e);
+                throw new OverflowException(SR.Overflow_SByte, e);
             }
 
-            // We need this check here since we don't allow signs to specified in hex numbers. So we fixup the result
-            // for negative numbers
             if ((style & NumberStyles.AllowHexSpecifier) != 0)
             { // We are parsing a hexadecimal number
-                if ((i < 0) || (i > UInt16.MaxValue))
+                if ((i < 0) || i > Byte.MaxValue)
                 {
-                    throw new OverflowException(SR.Overflow_Int16);
+                    throw new OverflowException(SR.Overflow_SByte);
                 }
-                return (short)i;
+                return (sbyte)i;
             }
 
-            if (i < MinValue || i > MaxValue) throw new OverflowException(SR.Overflow_Int16);
-            return (short)i;
+            if (i < MinValue || i > MaxValue) throw new OverflowException(SR.Overflow_SByte);
+            return (sbyte)i;
         }
 
-        public static bool TryParse(String s, out Int16 result)
+        [CLSCompliant(false)]
+        public static bool TryParse(String s, out SByte result)
         {
             return TryParse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo, out result);
         }
 
-        public static bool TryParse(String s, NumberStyles style, IFormatProvider provider, out Int16 result)
+        [CLSCompliant(false)]
+        public static bool TryParse(String s, NumberStyles style, IFormatProvider provider, out SByte result)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
             return TryParse(s, style, NumberFormatInfo.GetInstance(provider), out result);
         }
 
-        private static bool TryParse(String s, NumberStyles style, NumberFormatInfo info, out Int16 result)
+        private static bool TryParse(String s, NumberStyles style, NumberFormatInfo info, out SByte result)
         {
             result = 0;
             int i;
@@ -184,15 +185,13 @@ namespace System
                 return false;
             }
 
-            // We need this check here since we don't allow signs to specified in hex numbers. So we fixup the result
-            // for negative numbers
             if ((style & NumberStyles.AllowHexSpecifier) != 0)
             { // We are parsing a hexadecimal number
-                if ((i < 0) || i > UInt16.MaxValue)
+                if ((i < 0) || i > Byte.MaxValue)
                 {
                     return false;
                 }
-                result = (Int16)i;
+                result = (sbyte)i;
                 return true;
             }
 
@@ -200,7 +199,7 @@ namespace System
             {
                 return false;
             }
-            result = (Int16)i;
+            result = (sbyte)i;
             return true;
         }
 
@@ -210,7 +209,7 @@ namespace System
 
         public TypeCode GetTypeCode()
         {
-            return TypeCode.Int16;
+            return TypeCode.SByte;
         }
 
 
@@ -226,7 +225,7 @@ namespace System
 
         sbyte IConvertible.ToSByte(IFormatProvider provider)
         {
-            return Convert.ToSByte(m_value);
+            return m_value;
         }
 
         byte IConvertible.ToByte(IFormatProvider provider)
@@ -236,7 +235,7 @@ namespace System
 
         short IConvertible.ToInt16(IFormatProvider provider)
         {
-            return m_value;
+            return Convert.ToInt16(m_value);
         }
 
         ushort IConvertible.ToUInt16(IFormatProvider provider)
@@ -246,7 +245,7 @@ namespace System
 
         int IConvertible.ToInt32(IFormatProvider provider)
         {
-            return Convert.ToInt32(m_value);
+            return m_value;
         }
 
         uint IConvertible.ToUInt32(IFormatProvider provider)
@@ -281,7 +280,7 @@ namespace System
 
         DateTime IConvertible.ToDateTime(IFormatProvider provider)
         {
-            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "Int16", "DateTime"));
+            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "SByte", "DateTime"));
         }
 
         Object IConvertible.ToType(Type type, IFormatProvider provider)
