+++ src/mscorlib/shared/System/Single.cs
@@ -11,11 +11,9 @@
 **
 ===========================================================*/
 
-using System;
 using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Runtime.CompilerServices;
-using System.Runtime.ConstrainedExecution;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
