+++ src/mscorlib/shared/System/UInt16.cs
@@ -5,36 +5,37 @@
 /*============================================================
 **
 **
-** Purpose: This class will encapsulate an unsigned long and 
-**          provide an Object representation of it.
+** Purpose: This class will encapsulate a short and provide an
+**          Object representation of it.
 **
 ** 
 ===========================================================*/
 
-using System.Globalization;
-using System;
-using System.Runtime.InteropServices;
 using System.Diagnostics.Contracts;
+using System.Globalization;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
 
 namespace System
 {
-    // Wrapper for unsigned 64 bit integers.
     [Serializable]
-    [CLSCompliant(false), System.Runtime.InteropServices.StructLayout(LayoutKind.Sequential)]
-    [System.Runtime.CompilerServices.TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")] 
-    public struct UInt64 : IComparable, IFormattable, IConvertible
-            , IComparable<UInt64>, IEquatable<UInt64>
+    [CLSCompliant(false)]
+    [StructLayout(LayoutKind.Sequential)]
+    [TypeForwardedFrom("mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
+    public struct UInt16 : IComparable, IConvertible, IFormattable, IComparable<UInt16>, IEquatable<UInt16>
     {
-        private ulong m_value; // Do not rename (binary serialization)
+        private ushort m_value; // Do not rename (binary serialization)
 
-        public const ulong MaxValue = (ulong)0xffffffffffffffffL;
-        public const ulong MinValue = 0x0;
+        public const ushort MaxValue = (ushort)0xFFFF;
+        public const ushort MinValue = 0;
+
 
         // Compares this object to another object, returning an integer that
         // indicates the relationship. 
         // Returns a value less than zero if this  object
         // null is considered to be less than any instance.
-        // If object is not of type UInt64, this method throws an ArgumentException.
+        // If object is not of type UInt16, this method throws an ArgumentException.
         // 
         public int CompareTo(Object value)
         {
@@ -42,109 +43,135 @@ namespace System
             {
                 return 1;
             }
-            if (value is UInt64)
+            if (value is UInt16)
             {
-                // Need to use compare because subtraction will wrap
-                // to positive for very large neg numbers, etc.
-                ulong i = (ulong)value;
-                if (m_value < i) return -1;
-                if (m_value > i) return 1;
-                return 0;
+                return ((int)m_value - (int)(((UInt16)value).m_value));
             }
-            throw new ArgumentException(SR.Arg_MustBeUInt64);
+            throw new ArgumentException(SR.Arg_MustBeUInt16);
         }
 
-        public int CompareTo(UInt64 value)
+        public int CompareTo(UInt16 value)
         {
-            // Need to use compare because subtraction will wrap
-            // to positive for very large neg numbers, etc.
-            if (m_value < value) return -1;
-            if (m_value > value) return 1;
-            return 0;
+            return ((int)m_value - (int)value);
         }
 
         public override bool Equals(Object obj)
         {
-            if (!(obj is UInt64))
+            if (!(obj is UInt16))
             {
                 return false;
             }
-            return m_value == ((UInt64)obj).m_value;
+            return m_value == ((UInt16)obj).m_value;
         }
 
         [System.Runtime.Versioning.NonVersionable]
-        public bool Equals(UInt64 obj)
+        public bool Equals(UInt16 obj)
         {
             return m_value == obj;
         }
 
-        // The value of the lower 32 bits XORed with the uppper 32 bits.
+        // Returns a HashCode for the UInt16
         public override int GetHashCode()
         {
-            return ((int)m_value) ^ (int)(m_value >> 32);
+            return (int)m_value;
         }
 
+        // Converts the current value to a String in base-10 with no extra padding.
         public override String ToString()
         {
             Contract.Ensures(Contract.Result<String>() != null);
-            return Number.FormatUInt64(m_value, null, NumberFormatInfo.CurrentInfo);
+            return Number.FormatUInt32(m_value, null, NumberFormatInfo.CurrentInfo);
         }
 
         public String ToString(IFormatProvider provider)
         {
             Contract.Ensures(Contract.Result<String>() != null);
-            return Number.FormatUInt64(m_value, null, NumberFormatInfo.GetInstance(provider));
+            return Number.FormatUInt32(m_value, null, NumberFormatInfo.GetInstance(provider));
         }
+
 
         public String ToString(String format)
         {
             Contract.Ensures(Contract.Result<String>() != null);
-            return Number.FormatUInt64(m_value, format, NumberFormatInfo.CurrentInfo);
+            return Number.FormatUInt32(m_value, format, NumberFormatInfo.CurrentInfo);
         }
 
         public String ToString(String format, IFormatProvider provider)
         {
             Contract.Ensures(Contract.Result<String>() != null);
-            return Number.FormatUInt64(m_value, format, NumberFormatInfo.GetInstance(provider));
+            return Number.FormatUInt32(m_value, format, NumberFormatInfo.GetInstance(provider));
         }
 
         [CLSCompliant(false)]
-        public static ulong Parse(String s)
+        public static ushort Parse(String s)
         {
-            return Number.ParseUInt64(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo);
+            return Parse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo);
         }
 
         [CLSCompliant(false)]
-        public static ulong Parse(String s, NumberStyles style)
-        {
-            NumberFormatInfo.ValidateParseStyleInteger(style);
-            return Number.ParseUInt64(s, style, NumberFormatInfo.CurrentInfo);
-        }
-
-        [CLSCompliant(false)]
-        public static ulong Parse(string s, IFormatProvider provider)
-        {
-            return Number.ParseUInt64(s, NumberStyles.Integer, NumberFormatInfo.GetInstance(provider));
-        }
-
-        [CLSCompliant(false)]
-        public static ulong Parse(String s, NumberStyles style, IFormatProvider provider)
+        public static ushort Parse(String s, NumberStyles style)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
-            return Number.ParseUInt64(s, style, NumberFormatInfo.GetInstance(provider));
+            return Parse(s, style, NumberFormatInfo.CurrentInfo);
         }
 
+
         [CLSCompliant(false)]
-        public static Boolean TryParse(String s, out UInt64 result)
+        public static ushort Parse(String s, IFormatProvider provider)
         {
-            return Number.TryParseUInt64(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo, out result);
+            return Parse(s, NumberStyles.Integer, NumberFormatInfo.GetInstance(provider));
         }
 
         [CLSCompliant(false)]
-        public static Boolean TryParse(String s, NumberStyles style, IFormatProvider provider, out UInt64 result)
+        public static ushort Parse(String s, NumberStyles style, IFormatProvider provider)
         {
             NumberFormatInfo.ValidateParseStyleInteger(style);
-            return Number.TryParseUInt64(s, style, NumberFormatInfo.GetInstance(provider), out result);
+            return Parse(s, style, NumberFormatInfo.GetInstance(provider));
+        }
+
+        private static ushort Parse(String s, NumberStyles style, NumberFormatInfo info)
+        {
+            uint i = 0;
+            try
+            {
+                i = Number.ParseUInt32(s, style, info);
+            }
+            catch (OverflowException e)
+            {
+                throw new OverflowException(SR.Overflow_UInt16, e);
+            }
+
+            if (i > MaxValue) throw new OverflowException(SR.Overflow_UInt16);
+            return (ushort)i;
+        }
+
+        [CLSCompliant(false)]
+        public static bool TryParse(String s, out UInt16 result)
+        {
+            return TryParse(s, NumberStyles.Integer, NumberFormatInfo.CurrentInfo, out result);
+        }
+
+        [CLSCompliant(false)]
+        public static bool TryParse(String s, NumberStyles style, IFormatProvider provider, out UInt16 result)
+        {
+            NumberFormatInfo.ValidateParseStyleInteger(style);
+            return TryParse(s, style, NumberFormatInfo.GetInstance(provider), out result);
+        }
+
+        private static bool TryParse(String s, NumberStyles style, NumberFormatInfo info, out UInt16 result)
+        {
+            result = 0;
+            UInt32 i;
+            if (!Number.TryParseUInt32(s, style, info, out i))
+            {
+                return false;
+            }
+            if (i > MaxValue)
+            {
+                return false;
+            }
+            result = (UInt16)i;
+            return true;
         }
 
         //
@@ -153,7 +180,7 @@ namespace System
 
         public TypeCode GetTypeCode()
         {
-            return TypeCode.UInt64;
+            return TypeCode.UInt16;
         }
 
         bool IConvertible.ToBoolean(IFormatProvider provider)
@@ -183,7 +210,7 @@ namespace System
 
         ushort IConvertible.ToUInt16(IFormatProvider provider)
         {
-            return Convert.ToUInt16(m_value);
+            return m_value;
         }
 
         int IConvertible.ToInt32(IFormatProvider provider)
@@ -203,7 +230,7 @@ namespace System
 
         ulong IConvertible.ToUInt64(IFormatProvider provider)
         {
-            return m_value;
+            return Convert.ToUInt64(m_value);
         }
 
         float IConvertible.ToSingle(IFormatProvider provider)
@@ -223,7 +250,7 @@ namespace System
 
         DateTime IConvertible.ToDateTime(IFormatProvider provider)
         {
-            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "UInt64", "DateTime"));
+            throw new InvalidCastException(SR.Format(SR.InvalidCast_FromTo, "UInt16", "DateTime"));
         }
 
         Object IConvertible.ToType(Type type, IFormatProvider provider)
